Move icon button to shared scope
