+++ Tireless/Auth/Sign in/View/SignInView.swift
@@ -118,60 +118,6 @@ struct SignInView: View {
     }
 }
 
-struct IconButton: View {
-    enum ImageName {
-        case custom(String)
-        case system(String)
-    }
-    
-    private let action: () -> Void
-    private let imageName: ImageName
-    private let size: CGSize
-
-    init(imageName: ImageName, size: CGSize, action: @escaping () -> Void) {
-        self.imageName = imageName
-        self.size = size
-        self.action = action
-    }
-    
-    var body: some View {
-        Button(action: {
-            self.action()
-        }) {
-            ZStack {
-                Color.white
-                    .frame(width: size.width, height: size.height)
-                    .cornerRadius(8)
-                    .shadow(color: Color.gray.opacity(0.3), radius: 5, x: 0, y: 2)
-                
-                switch imageName {
-                case let .custom(name):
-                    Image(name)
-                        .resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .frame(width: 28, height: 28)
-                    
-                case let .system(name):
-                    Image(systemName: name)
-                        .resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .frame(width: 28, height: 28)
-                }
-                
-            }
-        }
-        .buttonStyle(GrowingButton())
-    }
-}
-
-struct GrowingButton: ButtonStyle {
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .scaleEffect(configuration.isPressed ? 1.2 : 1)
-            .animation(.easeOut(duration: 0.2), value: configuration.isPressed)
-    }
-}
-
 struct SignInView_Previews: PreviewProvider {
     static var previews: some View {
         SignInView(viewModel: SignInViewModel()) {}
